refactor(calculator): Lotto 예외처리
